ISA/requirements.txt
asgiref==3.6.0
certifi==2021.10.8
Django==4.1.5
djangorestframework==3.15.0
django-cors-headers
psycopg2-binary
sqlparse==0.4.3
gunicorn
python-dotenv==1.0.1